requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile .\pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.7.9
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.176.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via chatbots (./pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
narwhals==1.46.0
    # via altair
numpy==2.3.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.1
    # via streamlit
pillow==11.3.0
    # via streamlit
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==20.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via google-generativeai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via chatbots (./pyproject.toml)
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   google-api-core
    #   streamlit
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.46.1
    # via chatbots (./pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.14.1
    # via
    #   altair
    #   google-generativeai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
